+++ fiddles/angular2-cli/fiddle-0039-WebNotificationsHW/src/app/notification-queue/notification-queue.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-notification-queue',
+  templateUrl: './notification-queue.component.html',
+  styleUrls: ['./notification-queue.component.css']
+})
+export class NotificationQueueComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}